+++ src/tf/parse_gather.cpp
@@ -1,0 +1,27 @@
+#include <migraphx/tf/op_parser.hpp>
+#include <migraphx/tf/tf_parser.hpp>
+#include <migraphx/ranges.hpp>
+#include <migraphx/instruction.hpp>
+#include <migraphx/make_op.hpp>
+
+namespace migraphx {
+inline namespace MIGRAPHX_INLINE_NS {
+namespace tf {
+
+struct parse_gather : op_parser<parse_gather>
+{
+    std::vector<op_desc> operators() const { return {{"GatherV2"}}; }
+
+    instruction_ref parse(const op_desc& /*opd*/,
+                          const tf_parser& /*parser*/,
+                          const tf_parser::node_info& info,
+                          std::vector<instruction_ref> args) const
+    {
+        int axis = args[2]->eval().at<int32_t>();
+        return info.add_instruction(make_op("gather", {{"axis", axis}}), {args[0], args[1]});
+    }
+};
+
+} // namespace tf
+} // namespace MIGRAPHX_INLINE_NS
+} // namespace migraphx
